% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_rawdataQC_stack.R
\name{Plot_rawdataQC_stack}
\alias{Plot_rawdataQC_stack}
\title{Title}
\usage{
Plot_rawdataQC_stack(
  data = data,
  xlab = xlab,
  var1 = "exonic",
  var2 = "intronic",
  var3 = "intergenic",
  limit_var1 = 0.4,
  limit_var2 = 0.3,
  limit_var3 = 0,
  changelabname = F,
  xlabname = xlabname,
  ylabname = ylabname,
  var1labname = "Exonic rate",
  var2labname = "Intronic rate",
  var3labname = "Intergenic rate",
  var3_color = "#66C2A5",
  var2_color = "#FC8D62",
  var1_color = "#8DA0CA",
  title = F,
  titlename = titlename,
  baralpha = 0.5,
  barwidth = 0.8,
  bar_pass_color = "#ACACAC",
  bar_fail_color = "#D32F2E",
  facet = F,
  facet_label = "batch",
  showerror = F,
  showsample = F,
  axisx_size = 12,
  axisy_size = 16,
  text_size = 20,
  axis_title_size = 16
)
}
\arguments{
\item{axis_title_size}{}
}
\value{

}
\description{
Title
}
