% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_rawdataQC_standard.R
\name{Plot_rawdataQC_standard}
\alias{Plot_rawdataQC_standard}
\title{Title}
\usage{
Plot_rawdataQC_standard(
  data = data,
  xlab = xlab,
  ylab = ylab,
  changelabname = F,
  xlabname = xlabname,
  ylabname = ylabname,
  title = F,
  titlename = titlename,
  baralpha = 0.5,
  barwidth = 0.8,
  bar_pass_color = "#ACACAC",
  bar_fail_color = "#D32F2E",
  facet = F,
  facet_label = "batch",
  showsampletype = F,
  sampletype_label = "type",
  showerror = F,
  showsample = F,
  limit = limit,
  limit_color = "red",
  limit_size = 1,
  axisx_size = 12,
  axisy_size = 16,
  text_size = 20,
  axis_title_size = 16,
  showerrordetail = T,
  errordetail_color = "black",
  numberdig = 3
)
}
\arguments{
\item{numberdig}{}
}
\value{

}
\description{
Title
}
